注文一覧・詳細画面

+++ app/controllers/public/orders_controller.rb
@@ -30,17 +30,21 @@ class Public::OrdersController < ApplicationController
   def create
     @cart_items = current_customer.cart_items.all
     @order = Order.new(order_params)
-    @order.save
-    @cart_items.each do |cart_item|
-      @order_detail = OrderDetail.new
-      @order_detail.purchase_price = cart_item.item.with_tax_price
-      @order_detail.amount = cart_item.amount
-      @order_detail.order_id = order.id
-      @order_detail.item_id = cart_item.item.id
-      @order_detail.save
+    if @order.save
+      @cart_items.each do |cart_item|
+        @order_detail = OrderDetail.new
+        @order_detail.purchase_price = cart_item.item.with_tax_price
+        @order_detail.amount = cart_item.amount
+        @order_detail.order_id = order.id
+        @order_detail.item_id = cart_item.item.id
+        @order_detail.save
+      end
+      current_customer.cart_items.destroy_all
+      redirect_to public_order_complete_path
+    else
+      flash[:notice] = "注文が確定できませんでした。もう一度やり直してください。"
+      redirect_to public_cart_items_path
     end
-    current_customer.cart_items.destroy_all
-    redirect_to public_order_complete_path
   end
 
   def complete
@@ -51,7 +55,7 @@ class Public::OrdersController < ApplicationController
   end
 
   def show
-    # @order = Order.find(params[:id])
+    @order = Order.find(params[:id])
     # @order = Order.where(customer_id: current_customer_id)
   end
